Fix #1: repeated subdivision should work
`truncateLexHigher` was rounding the 'larger' string to 1+length of the
shorter one to save computation, but that can't be done reliably without
ensuring that the truncated string doesn't end in the 0 symbol, which
requires knowledge of the symbol table.

Rather than make this function more complicated, I just removed this
micro-optimization.

The test shows it subdividing the a-z interval fifty times, ending in
something like `aaaaaaaaag` :P